Moved SyntaxTree and SyntaxTreeExtensions
