+++ src/Reline.Compilation/Syntax/SyntaxTree.cs
@@ -1,7 +1,7 @@
 ﻿using Reline.Compilation.Syntax.Nodes;
 using Reline.Compilation.Diagnostics;
 
-namespace Reline.Compilation.Parsing;
+namespace Reline.Compilation.Syntax;
 
 public sealed class SyntaxTree {
 
@@ -10,7 +10,7 @@ public sealed class SyntaxTree {
 
 
 
-	public SyntaxTree(ProgramSyntax root, ImmutableArray<Diagnostic> diagnostics) {
+	private SyntaxTree(ProgramSyntax root, ImmutableArray<Diagnostic> diagnostics) {
 		Root = root;
 		Diagnostics = diagnostics;
 	}
